Header: spin off HeaderSearchForm
